Commitando projeto com problemas de conectar com o BD postgresql

+++ src/main/java/br/com/prematricula/ws/repository/AlunoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.prematricula.ws.model.Aluno;
 
-public interface AlunoRepository extends JpaRepository<Aluno,Long> {
+public interface AlunoRepository extends JpaRepository<Aluno,Integer> {
 
 }
